requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Other Packages
python-dotenv

# Flask Framework
Flask==2.2

# Flask Packages
Flask-Login==0.6
Flask-Migrate==3.1
# Flask-Script==2.0.5
Flask-SQLAlchemy==3.0
Flask_RESTful==0.3.9
# Flask-WTF==0.14.2
# Flask-User==1.0.1.5

# Automated tests
# pytest==3.0.5
# pytest-cov==2.4.0

# flask_bcrypt==3.2.0
